requirements.txt
# Core Flask
Flask==3.1.0
flask-cors==6.0.1
Werkzeug==3.1.3

# Database
psycopg2-binary==2.9.10

# HTTP Client
requests==2.32.3
urllib3==1.26.20

# Cache
redis==5.0.1

# Environment
python-dotenv==1.0.1

# Utilities
click==8.1.8

# Production Server
gunicorn==23.0.0

# Development
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.0
flake8==7.0.0
